requirements.txt
django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
boto3==1.38.43
pillow==11.2.1
python-decouple==3.8
django-storages==1.14.6
dj-database-url==3.0.0

